refactor: Move constants from mapper to volume property
