+++ Sources/Rendering/Core/VolumeMapper/Constants.d.ts
@@ -1,21 +1,8 @@
-export declare enum BlendMode {
-  COMPOSITE_BLEND = 0,
-  MAXIMUM_INTENSITY_BLEND = 1,
-  MINIMUM_INTENSITY_BLEND = 2,
-  AVERAGE_INTENSITY_BLEND = 3,
-  ADDITIVE_INTENSITY_BLEND = 4,
-  RADON_TRANSFORM_BLEND = 5,
-  LABELMAP_EDGE_PROJECTION_BLEND = 6,
-}
+// Don't use the constants from this file
 
-export declare enum FilterMode {
-  OFF = 0,
-  NORMALIZED = 1,
-  RAW = 2,
-}
-
-declare const _default: {
-  BlendMode: typeof BlendMode;
-  FilterMode: typeof FilterMode;
-};
-export default _default;
+// Prefer constants from volume property:
+export {
+  default,
+  BlendMode,
+  FilterMode,
+} from '../VolumeProperty/Constants';
